test(github): move validation and content checks out of the monolith
Second half of splitting the 1,390-line tests/test_new_pr.py. Removes the
original file and lands the two remaining modules:

  tests/test_new_pr_validations.py     485  pre-creation validation pipeline
  tests/test_new_pr_content_checks.py  432  dash, escaped newline, output codec

Together with the harness commit, every module is under the 500-line
limit and no test was renamed, weakened, or dropped: 86 tests before, 86
after, plus the 6 warning-reporting tests added earlier in this branch.
All 92 pass.

Refs #4764

Copilot-Session: 2005b9d4-c29c-460c-8b98-878535ec8604
